third_party/blink/web_tests/external/wpt/sanitizer-api/sanitizer-sanitize.https.tentative-expected.txt

This is a testharness.js-based test.
Found 56 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] SanitizerAPI with config: test html without close tag with removeElements list ['div'], sanitize from document function for <body>
  assert_equals: expected "<p>bla</p>" but got "<div>test</div><p>bla</p>"
[FAIL] SanitizerAPI with config: allow custom elements, sanitize from document function for <body>
  assert_equals: expected "testbla" but got "bla"
[FAIL] SanitizerAPI with config: allow custom elements with allow elements, sanitize from document function for <body>
  assert_equals: expected "<custom-element>test</custom-element>bla" but got "bla"
[FAIL] SanitizerAPI with config: removeElements list ["test-element", "i"]}, sanitize from document function for <body>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: removeElements list ["dl", "p"]}, sanitize from document function for <body>
  assert_equals: expected "<div>balabala<i>i</i></div>" but got "<div>balabala<i>i</i><p>t</p></div>"
[FAIL] SanitizerAPI with config: elements list ["p"], sanitize from document function for <body>
  assert_equals: expected "testptt<p>div</p>" but got "<div>test<div>p</div>tt<p>div</p></div>"
[FAIL] SanitizerAPI with config: elements list has no influence to removeElements, sanitize from document function for <body>
  assert_equals: expected "bla" but got "<div>test</div><p>bla</p>"
[FAIL] SanitizerAPI with config: removeAttributes list ["id"] with id attribute, sanitize from document function for <body>
  assert_equals: expected "<p>Click.</p>" but got "<p id=\\"test\\">Click.</p>"
[FAIL] SanitizerAPI with config: elements list with <p> attributes: ["title"] and div attributes: ["id"] lists, sanitize from document function for <body>
  assert_equals: expected "<p title=\\"p\\">P</p><div id=\\"div\\">DIV</div>" but got "<p id=\\"p\\" title=\\"p\\">P</p><div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <p> removeAttributes: ["title"]  and div removeAttributes: ["id"] lists, sanitize from document function for <body>
  assert_equals: expected "<p id=\\"p\\">P</p><div title=\\"div\\">DIV</div>" but got "<p id=\\"p\\" title=\\"p\\">P</p><div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <div> attributes: ["id"] and removeAttributes: ["id"] lists, sanitize from document function for <body>
  assert_equals: expected "<div>DIV</div>" but got "<div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <div> attributes: ["id", "title"] does not override empty attributes: [] list, sanitize from document function for <body>
  assert_equals: expected "<div>DIV</div>" but got "<div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <div> attributes: ["id", "title"] does not override removeAttributes: ["id", "title"] list, sanitize from document function for <body>
  assert_equals: expected "<div>DIV</div>" but got "<div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <div> removeAttributes: ["id", "title"] is effective even with attributes: ["id", "title"] list, sanitize from document function for <body>
  assert_equals: expected "<div>DIV</div>" but got "<div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: attributes list has no influence to removeAttributes list, sanitize from document function for <body>
  assert_equals: expected "<p>Click.</p>" but got "<p style=\\"color: black\\">Click.</p>"
[FAIL] SanitizerAPI with config: Template element, sanitize from document function for <body>
  assert_equals: expected "<template><div>hello</div></template>" but got ""
[FAIL] SanitizerAPI with config: HTML with comments; comments, sanitize from document function for <body>
  assert_equals: expected "<p>Some text</p><!-- 1 --><!-- 2 --><p>Some more text</p>" but got "<p>Some text</p><p>Some more text</p>"
[FAIL] SanitizerAPI with config: HTML with comments deeper in the tree, comments, sanitize from document function for <body>
  assert_equals: expected "<p>comment<!-- hello -->in<!-- </p> -->text</p>" but got "<p>commentintext</p>"
[FAIL] SanitizerAPI with config: removeElements list ["i", "dl"]}, sanitize from document function for <body>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<dl>test</dl></div>"
[FAIL] SanitizerAPI with config: removeElements list ["i", "dl"]} with uppercase HTML, sanitize from document function for <body>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<dl>test</dl></div>"
[FAIL] SanitizerAPI with config: removeAttributes list ["id"] with ID attribute, sanitize from document function for <body>
  assert_equals: expected "<p>Click.</p>" but got "<p id=\\"test\\">Click.</p>"
[FAIL] SanitizerAPI with config: removeElements with unknown elements and without unknownMarkup, sanitize from document function for <body>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: replaceWithChildrenElements with unknown elements and without unknownMarkup, sanitize from document function for <body>
  assert_equals: expected "<div>balabalatest</div>" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: elements with unknown elements and without unknownMarkup, sanitize from document function for <body>
  assert_equals: expected "testptt<p>div</p>" but got "<div>test<div>p</div>tt<p>div</p></div>"
[FAIL] SanitizerAPI with config: removeElements with unknown elements and with unknownMarkup, sanitize from document function for <body>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: replaceWithChildrenElements with unknown elements and with unknownMarkup, sanitize from document function for <body>
  assert_equals: expected "<div>balabalatest</div>t" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: elements with unknown elements and with unknownMarkup, sanitize from document function for <body>
  assert_equals: expected "testptt<p>div</p><test>test</test>" but got "<div>test<div>p</div>tt<p>div</p></div>"
[FAIL] SanitizerAPI with config: attributes: unknown attributes and with unknownMarkup, sanitize from document function for <body>
  assert_equals: expected "<div hello=\\"1\\" world=\\"2\\"><b hello=\\"3\\" world=\\"4\\"></b></div>" but got "<div><b></b></div>"
[FAIL] SanitizerAPI with config: test html without close tag with removeElements list ['div'], sanitize from document fragment function for <template>
  assert_equals: expected "<p>bla</p>" but got "<div>test</div><p>bla</p>"
[FAIL] SanitizerAPI with config: allow custom elements, sanitize from document fragment function for <template>
  assert_equals: expected "testbla" but got "bla"
[FAIL] SanitizerAPI with config: allow custom elements with allow elements, sanitize from document fragment function for <template>
  assert_equals: expected "<custom-element>test</custom-element>bla" but got "bla"
[FAIL] SanitizerAPI with config: removeElements list ["test-element", "i"]}, sanitize from document fragment function for <template>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: removeElements list ["dl", "p"]}, sanitize from document fragment function for <template>
  assert_equals: expected "<div>balabala<i>i</i></div>" but got "<div>balabala<i>i</i><p>t</p></div>"
[FAIL] SanitizerAPI with config: elements list ["p"], sanitize from document fragment function for <template>
  assert_equals: expected "testptt<p>div</p>" but got "<div>test<div>p</div>tt<p>div</p></div>"
[FAIL] SanitizerAPI with config: elements list has no influence to removeElements, sanitize from document fragment function for <template>
  assert_equals: expected "bla" but got "<div>test</div><p>bla</p>"
[FAIL] SanitizerAPI with config: removeAttributes list ["id"] with id attribute, sanitize from document fragment function for <template>
  assert_equals: expected "<p>Click.</p>" but got "<p id=\\"test\\">Click.</p>"
[FAIL] SanitizerAPI with config: elements list with <p> attributes: ["title"] and div attributes: ["id"] lists, sanitize from document fragment function for <template>
  assert_equals: expected "<p title=\\"p\\">P</p><div id=\\"div\\">DIV</div>" but got "<p id=\\"p\\" title=\\"p\\">P</p><div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <p> removeAttributes: ["title"]  and div removeAttributes: ["id"] lists, sanitize from document fragment function for <template>
  assert_equals: expected "<p id=\\"p\\">P</p><div title=\\"div\\">DIV</div>" but got "<p id=\\"p\\" title=\\"p\\">P</p><div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <div> attributes: ["id"] and removeAttributes: ["id"] lists, sanitize from document fragment function for <template>
  assert_equals: expected "<div>DIV</div>" but got "<div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <div> attributes: ["id", "title"] does not override empty attributes: [] list, sanitize from document fragment function for <template>
  assert_equals: expected "<div>DIV</div>" but got "<div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <div> attributes: ["id", "title"] does not override removeAttributes: ["id", "title"] list, sanitize from document fragment function for <template>
  assert_equals: expected "<div>DIV</div>" but got "<div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: elements list with <div> removeAttributes: ["id", "title"] is effective even with attributes: ["id", "title"] list, sanitize from document fragment function for <template>
  assert_equals: expected "<div>DIV</div>" but got "<div id=\\"div\\" title=\\"div\\">DIV</div>"
[FAIL] SanitizerAPI with config: attributes list has no influence to removeAttributes list, sanitize from document fragment function for <template>
  assert_equals: expected "<p>Click.</p>" but got "<p style=\\"color: black\\">Click.</p>"
[FAIL] SanitizerAPI with config: Template element, sanitize from document fragment function for <template>
  assert_equals: expected "<template><div>hello</div></template>" but got ""
[FAIL] SanitizerAPI with config: HTML with comments; comments, sanitize from document fragment function for <template>
  assert_equals: expected "<p>Some text</p><!-- 1 --><!-- 2 --><p>Some more text</p>" but got "<p>Some text</p><p>Some more text</p>"
[FAIL] SanitizerAPI with config: HTML with comments deeper in the tree, comments, sanitize from document fragment function for <template>
  assert_equals: expected "<p>comment<!-- hello -->in<!-- </p> -->text</p>" but got "<p>commentintext</p>"
[FAIL] SanitizerAPI with config: removeElements list ["i", "dl"]}, sanitize from document fragment function for <template>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<dl>test</dl></div>"
[FAIL] SanitizerAPI with config: removeElements list ["i", "dl"]} with uppercase HTML, sanitize from document fragment function for <template>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<dl>test</dl></div>"
[FAIL] SanitizerAPI with config: removeAttributes list ["id"] with ID attribute, sanitize from document fragment function for <template>
  assert_equals: expected "<p>Click.</p>" but got "<p id=\\"test\\">Click.</p>"
[FAIL] SanitizerAPI with config: removeElements with unknown elements and without unknownMarkup, sanitize from document fragment function for <template>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: replaceWithChildrenElements with unknown elements and without unknownMarkup, sanitize from document fragment function for <template>
  assert_equals: expected "<div>balabalatest</div>" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: elements with unknown elements and without unknownMarkup, sanitize from document fragment function for <template>
  assert_equals: expected "testptt<p>div</p>" but got "<div>test<div>p</div>tt<p>div</p></div>"
[FAIL] SanitizerAPI with config: removeElements with unknown elements and with unknownMarkup, sanitize from document fragment function for <template>
  assert_equals: expected "<div>balabala</div>" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: replaceWithChildrenElements with unknown elements and with unknownMarkup, sanitize from document fragment function for <template>
  assert_equals: expected "<div>balabalatest</div>t" but got "<div>balabala<i>test</i></div>"
[FAIL] SanitizerAPI with config: elements with unknown elements and with unknownMarkup, sanitize from document fragment function for <template>
  assert_equals: expected "testptt<p>div</p><test>test</test>" but got "<div>test<div>p</div>tt<p>div</p></div>"
[FAIL] SanitizerAPI with config: attributes: unknown attributes and with unknownMarkup, sanitize from document fragment function for <template>
  assert_equals: expected "<div hello=\\"1\\" world=\\"2\\"><b hello=\\"3\\" world=\\"4\\"></b></div>" but got "<div><b></b></div>"
Harness: the test ran to completion.
